OF-2272: No longer use MUCRole in MUCUser
MUCRole and MUCUser are both going to exist in (clustered) caches. To avoid sending duplicated data, they should no longer exist as field values of each-other.

This commit removes the MUCRole collection that was maintained in MUCUser, replacing it with a collection of room names in which the MUCUser has some kind of MUCRole.

It is important that the content of MUCUser#roomNames and MUCRoom#ROOM_OCCUPANTS_CACHE (which are MUCRoles) are kept in sync. This commit attempts to enforce that by making methods that add or remove content from MUCRoom#ROOM_OCCUPANTS_CACHE (such as MUCRoom#removeOccupantRole(MUCRole)}) _uniquely_ responsible for maintaining the content of MUCUser's 'roomNames'.