OF-2224: Maintain minimal data for all occupants in the cluster
There's a need to maintain some data that can be used to 'clean up' after a clusternode leaves the cluster. Primarily, this is needed to send presence unavailable stanzas to locally connected occupants of MUC rooms for occupants that are connected on cluster nodes that are now no longer available. As a cluster-composition event causes a clustered cache to be reset/cleaned, this data needs to be maintained _outside_ of such a cache. This commit introduces an OccupantManager that maintains this data, which uses ClusterTasks to update data across the the cluster.

This code needs further work, as it currently assumes that the cluster is fully operational, but not yet used (population of OccupantManager data upon joining/leaving a cluster that's already in active use needs to be added).

The data maintained by the new OccupantManager is a neccesary duplicate of some data that is maintained in the cache that maintains all MUCRoom instances, for the reason described above. It, however, also is a duplcate of the data that was maintained by MUCUser instances (and the corresponding cache). To avoid the need to maintain additional data duplication, this commit removes MUCUser. The state that it maintained is now obtained from the new OccupantManger. Most of the functionality that it possessed is now merged with MultiUserChatServiceImpl.