Add diagnostics for RoutingTableImpl-maintained s2s connection state
Whenever a clustered cache is used, two supporting data structures are typically needed:
- one that can be used to 'repopulate' the cache after it flips between clustered and non-clustered, that holds local data.
- one that is used after a cluster breakage, that holds non-local data.

These three data structures (cache, local, non-local) have some degree of data duplication between them. It is therefor important that these data structures remain consistent.

This commit introduces a method that checks for consistency, reporting its detailed findings as text. A JSP page (that isn't linked in a menu, you need to know its URL) is also added, to easily look at the data in a running server.

There's much to be improved here (I should've used a better data structure for this reporting), and much to be added (basically, all other caches in Openfire), but this is a start, based on which I've already been able to squash some bugs.