Make AbstractCacheEntryListener work for cache entries not owned by one cluster node
Up until now, all cache entries that we've processed in this rewrite were 'owned' by one specific cluster node (eg: a client is connected to exactly one server, hence it's routing table cache entry can be linked to that one cluster node).

This, however, is not always the case. Components, for example, can connect to multiple server nodes. For the AbstractCacheEntryListener to work with these caches, cache updates (that can relate the cache entry to a different set of cluster nodes), need to be processed.