OF-2298: Reverse order of cluster event listeners
To be able to distinguish between the cause for an occupant to 'leave' a MUC room, it is important that we reverse the order in which the Cluster event listeners are fired.

Without this change, the SessionManager/RoutingTable would mark a user as unavailable. As a result of that, the MUC implementation would make the corresponding occupant leave the room, as if it sent a presence unavailable. By ensuring that the MUC-based cluster event listener is invoked first, the cause (a cluster breakage) for the occupant to become unavailable is known. This can be used to add status code 333 to the presence unavailable that is sent on behalf of the occupant that is dropping out.