Replaced term 'hostname' with 'domain' where appropriate in SessionManager
Older versions of Openfire used the term 'hostname' and 'domain' interchangebly. With clustering though, a single XMPP domain can consists of multiple servers, that each have their own hostname.

This commit replaces most references to the word 'hostname' with 'domain' in the implementation of SessionManager. No functional changes are intended.