OF-2289: Do not use Interner<String>
com.google.common.collect.Interner is not safe when used with Strings: as with String.intern(), an equal string value could end up being used as a mutex for unrelated code blocks (introducing the risk of deadlocks between code blocks that have no functional relation).

When an Interner instance is used with anohter immutable type, the generated mutexes are local to the Interner instance. This allows, for example, for the same JID value to be used with two different Interner instances, without the resulting mutexes influencing each-other.

This commit replaces all occurrances of Interner<String>.