OF-2325: Prevent endless loop when peer closes during TLS handshake
When a peer closes the connection while performing an initial TLS handshake, a problem occurs if the state of the handshake remains 'not complete': the code in TLSStreamHandler#start() will loop until the handshake is complete. Subsequent iterations will read nothing, which repeats indefinitely. This can drain the available CPU resources.

Instead, throwing an exception will force the invoker to give up.