OF-2322: Do not re-auth s2s for every queued stanza
When an outgoing s2s session is being established, stanzas addressed to the remote domain are cached. Only the first (and not all) queued packets should attempt to establish the connection.

The original code would attempt to establish a session for each queued stanza, which is handled somewhat gracefully in the sense that the existing session is used. However, an event is fired for each attempt. That event is picked up by an event listener that sends a new stanza to that domain, which causes recursion and a never-empty queue.

With this fix, only the first stanza (or the first stanza after a failure) attempts to establish the connection.