OF-2328: Implement optional Lock functionality for local locks
The clustered locks (based on Hazelcast) have some methods implemented that are not available in the non-clustered code.

The changes that have been applied to MUC's clustering implementation (OF-2220 for example) has introduced usage of these optional methods. This results in failing functionality when running Openfire in a non-clustered setup.

This commit adds an implementation for these optional methods.