OF-2321: Use DomainPair instead of 'remote domain'
A server-to-server connection is always established in context of a local and remote domain pair. The original code only defined the remote domain. This might introduce problems when two different local domains (eg: "example.org" and "conference.example.org") are establishing s2s at the same time.

Some time ago, the DomainPair class was introduced to represent this context. This commit replaces all usage of a singular 'domain' (or two distinct attributes for a local and remote domain) with a DomainPair instance.