OF-2321: Reworked OutgoingSessionPromise to improve thread safety
The existing implementation had synchronisation issues that allowed stanzas to be queued after the outgoing session was established (maybe keeping them in queue forever, effectively droppig them). Also, new stanzas wouldn't wait for queued stanzas to be delivered, causing stanzas to be sent out of order.

The changes in this commit should improve on that.

Sadly, the 'fast discard' option, which introduced a small delay before a failed s2s-connection could be retried, was proving to hard to work with during this refactoring. It has been dropped for now.