Shared groups: onlyGroup and recursion
When shared groups are configured to be shared with (users in) other groups, the corresponding value for the 'sharedRoster.showInRoster' property is 'onlyGroup'.

This property value has one of two meanings:
- the shared group is shared with (users in) the group itself
- the shared group is shared with (users in) a list of groups (potentially including itself).

A second property 'sharedRoster.groupList' is used to distinguish between the two: when its value is empty, then the shared group is shared with itself, otherwise, it's shared with the comma-separated list that is the value.

This commit more clearly defines this distinction, by changing the behavior of org.jivesoftware.openfire.group.AbstractGroupProvider#getVisibleGroupNames(java.lang.String) to return 'the group itself' when the property values indicate this behaviour.

When sharing groups with groups that themselves are shared with other groups, there's a chance for recursion to be desirable. Example:
- Users in SharedGroupA are visible to users of SharedGroupB
- Users in SharedGroupB are visible to users of SharedGroupC
With recursion disabled, users in SharedGroupC would see only the users of SharedGroupB.
With recursion enabled, users in SharedGroupC would see users of both SharedGroupA and SharedGroupB.

This commit adds a new property, abstractGroupProvider.shared.recursive, that controls this behavior.