OF-2411: Prevent deadlock in XmlProperties
OF-2377 introduced an updated implementation of the thread-safety module of XmlProperties (the old one wasn't fully thread-safe).

This new implementation causes a very specific problem at startup: When the initial attempt to migrate an XmlProperty is undertaken, TaskEngine is instantiated, which in turn tries to read/write properties (while under lock). This causes a deadlock.

To work around this issue, this commit replaces the usage of TaskEngine with a throw-away thread (that does not use any properties).