LGTM: ignore insecure-cookie when unsetting the cookie.
As long as we're exposing the admin console on a non-HTTPS endpoint (our port 9090), we will have a need to set cookies on an unsecure endpoint. This raises (rightfully) the LGTM 'insecure-cookie' alert, as Insecure cookies may be sent in cleartext, which makes them vulnerable to interception.

However, LGTM also raises this alert when we remove the cookie (which we do by setting an immediately-expired cookie with no data). The alert can be suppressed for that one.